PYIC-3506 Refactor code to separate criResponseService as a separate dependent service.
