+++ lambdas/process-async-cri-credential/build.gradle
@@ -17,7 +17,8 @@ dependencies {
 			"software.amazon.awssdk:dynamodb-enhanced:$rootProject.ext.dependencyVersions.dynamodbEnhanced",
 			project(":lib"),
 			project(":libs:vc-helper"),
-			project(":libs:cimit-service")
+			project(":libs:cimit-service"),
+			project(":libs:cri-response-service")
 	implementation project(path: ':libs:verifiable-credentials')
 
 	aspect "software.amazon.lambda:powertools-logging:$rootProject.ext.dependencyVersions.powertoolsLogging",
